Add a right-click edit/delete menu for both desktop and mobile! :D
